+++ packages/jsii-dotnet-runtime/src/Amazon.JSII.Runtime/Services/Runtime.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 
 namespace Amazon.JSII.Runtime.Services
 {
@@ -9,11 +10,22 @@ namespace Amazon.JSII.Runtime.Services
         public Runtime(INodeProcess nodeProcess)
         {
             _nodeProcess = nodeProcess ?? throw new ArgumentNullException(nameof(nodeProcess));
+            if (Environment.GetEnvironmentVariable("JSII_DEBUG") != null)
+            {
+                Task.Run(() => RedirectStandardError());
+            }
         }
 
         public string ReadResponse()
         {
-            return _nodeProcess.StandardOutput.ReadLine();
+            var response = _nodeProcess.StandardOutput.ReadLine();
+            if (string.IsNullOrEmpty(response))
+            {
+                var errorMessage = _nodeProcess.StandardError.ReadToEnd();
+                throw new JsiiException("Child process exited unexpectedly: " + errorMessage);
+            }
+
+            return response;
         }
 
         public void WriteRequest(string request)
@@ -30,6 +42,14 @@ namespace Amazon.JSII.Runtime.Services
 
             _nodeProcess.StandardInput.WriteLine(request);
             _nodeProcess.StandardInput.Flush();
+        }
+
+        private void RedirectStandardError()
+        {
+            while (true)
+            {
+                Console.Error.WriteLine(_nodeProcess.StandardError.ReadLine());
+            }
         }
     }
 }