+++ packages/jsii-pacmak/test/expected.jsii-calc/dotnet/Amazon.JSII.Tests.CalculatorPackageId/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using Amazon.JSII.Runtime.Deputy;
 
-[assembly: JsiiAssembly("jsii-calc", "0.7.11", "jsii-calc-0.7.11.tgz")]
+[assembly: JsiiAssembly("jsii-calc", "0.7.14", "jsii-calc-0.7.14.tgz")]