+++ packages/jsii-calc-base-of-base/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.14"></a>
+## [0.7.14](https://github.com/awslabs/jsii/compare/v0.7.13...v0.7.14) (2019-02-04)
+
+
+
+
+**Note:** Version bump only for package @scope/jsii-calc-base-of-base
+
+<a name="0.7.13"></a>
+## [0.7.13](https://github.com/awslabs/jsii/compare/v0.7.12...v0.7.13) (2019-01-03)
+
+
+
+
+**Note:** Version bump only for package @scope/jsii-calc-base-of-base
+
+<a name="0.7.12"></a>
+## [0.7.12](https://github.com/awslabs/jsii/compare/v0.7.11...v0.7.12) (2018-12-11)
+
+
+
+
+**Note:** Version bump only for package @scope/jsii-calc-base-of-base
+
 <a name="0.7.11"></a>
 ## [0.7.11](https://github.com/awslabs/jsii/compare/v0.7.10...v0.7.11) (2018-11-18)